Galleries&Pickers added
