+++ src/components/BasicInputs/MySlider/MySlider.tsx
